feat: Update Subject and Teacher Controllers to include SuperAdmin role in authorization
feat: Add Category property to Subject and Level entities in the database

feat: Implement DbInitializer for seeding roles and SuperAdmin user

fix: Correct constructor syntax in SubjectController

chore: Add ValidatorHelper for input validation

chore: Add migration for adding Category to Subjects and Levels

+++ Application/Models/Dto/StudentScoreDto.cs
@@ -9,7 +9,7 @@ namespace SchoolApp.Application.Models.Dto
     {
         public Guid StudentId { get; set; }
         public string? StudentName { get;  set; }
-        public int ContinuousAssessment { get; set; }
-        public int ExamScore { get; set; }
+        public double ContinuousAssessment { get; set; }
+        public double ExamScore { get; set; }
     }
 }
